Merging HTTP request functions with print & toggle
Wrote the previous work on to the new directory (team project(new)).
Changed some of complex.html entries according to user testing.
Assigned IDs to the html elements (some are missing for complex.html, waiting for discussion with client).
Uploading to the repo with educational license, on which we should work from now onwards.